1.Add StudentVo to query association table

+++ Springboot_Mybatis/src/main/java/com/example/entity/Student.java
@@ -25,7 +25,7 @@ public class Student {
 
     private Integer studentGender;
 
-    private Integer studentDepartmentId;
+    private Long studentDepartmentId;
